# services/cn-gateway/openapi/modea-cn-gateway.yaml
openapi: 3.0.3
info:
  title: Mode A CN Gateway
  version: "1.0.0"
  description: >
    Pseudonymous telemetry collector operating in Mainland China with strict Mode A DROP/SAFE enforcement.
servers:
  - url: https://cn-gateway.local
    description: mTLS-protected ingress via Nginx
components:
  securitySchemes:
    mTLS:
      type: mutualTLS
  schemas:
    Telemetry:
      $ref: ../src/validators/telemetry.schema.json
    IngestResponse:
      type: object
      required: [status, pseudoId, keyVersion]
      properties:
        status:
          type: string
          enum: [queued]
        pseudoId:
          type: string
          description: Pseudonymous identifier (22-24 characters).
        keyVersion:
          type: string
    BatchExport:
      type: object
      required:
        - batchId
        - count
        - records
      properties:
        batchId:
          type: string
        count:
          type: integer
        records:
          type: array
          items:
            type: object
            required: [pseudoId, seq, timestamp, metrics, keyVersion]
            properties:
              pseudoId:
                type: string
              seq:
                type: integer
              timestamp:
                type: string
                format: date-time
              keyVersion:
                type: string
              metrics:
                $ref: "#/components/schemas/Telemetry/properties/metrics"
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
paths:
  /ingest:
    post:
      security:
        - mTLS: []
      summary: Submit SAFE telemetry
      description: >
        Validates JSON payloads, enforces DROP/SAFE lists, pseudonymizes device identifiers, and schedules exports.
      parameters:
        - name: Idempotency-Key
          in: header
          required: false
          schema:
            type: string
          description: Stable identifier for 24h de-duplication.
        - name: X-Device-Signature
          in: header
          required: false
          schema:
            type: string
          description: Optional HMAC-SHA256 signature generated by the device key.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Telemetry"
      responses:
        "202":
          description: Accepted and queued.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IngestResponse"
        "401":
          description: Invalid device signature.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Replay detected for seq/timestamp or idempotency conflict.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "422":
          description: Validation or sanitization failure (DROP list hit, embedded identifiers, timestamp skew).
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "429":
          description: Per-device rate limit exceeded.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /export:
    post:
      security:
        - mTLS: []
      summary: Loopback endpoint to validate exporter batches.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BatchExport"
      responses:
        "202":
          description: Batch accepted.
        "400":
          description: Signature mismatch.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /admin/rotate-key:
    post:
      summary: Request CN KMS key rotation (simulated process).
      description: >
        Requires internal admin token plus TOTP. Records an audit trail and updates the in-memory KMS key version.
      parameters:
        - name: X-Admin-Token
          in: header
          required: true
          schema:
            type: string
        - name: X-Admin-TOTP
          in: header
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [keyVersion]
              properties:
                keyVersion:
                  type: string
      responses:
        "200":
          description: Rotation recorded.
        "400":
          description: Missing key version.
        "403":
          description: Unauthorized or invalid TOTP.
        "503":
          description: Rotation disabled (no admin credentials configured).
  /health:
    get:
      summary: Component health.
      responses:
        "200":
          description: Service healthy.
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                  version:
                    type: string
                  keyVersion:
                    type: string
                  exportEnabled:
                    type: boolean
  /metrics:
    get:
      summary: Prometheus exposition format.
      responses:
        "200":
          description: Metrics snapshot (disabled when METRICS_ENABLED=false).
          content:
            text/plain:
              schema:
                type: string
        "503":
          description: Metrics disabled.